CNI plugin based on sqlite3
Kubernetes container runtimes use CNI as the specification to provide
the network configuration to the Pod. Howecer, they only need two things
from the CNI plugin, the Pod IPs assigned and the creation of the Pod
network interface and its configuration.

However CNI plugins have some problems when Kubernetes advanced
networking capabilities are required:
- The especification is based on chaining and delegation without a
  strong schema, that puts the complexity on obscure fields and custom
  plugins, without providing a simple way of cooperating between
  different network plugins, you need to be part of the chain.
- The model is stateless, execute a chain of binaries piping information
  through them, that has multiple problems for scaling and running
  complex network applications, specially if those depend on multiple
  external factors.

Most of the existing kubernetes network CNI plugins resolve the
stateless problem by running a server in a unix socket, and creating a
small cni binary that just calls to the server, but this has the problem
that the cni binary now depends on a daemon that may or not may be
available.

Since Kindnet want to avoid that problem, instead of moving the
complexity to the daemon, we move it to the plugin in a reasonable way,
same as the host-local ipam plugin use files on the system to store the
allocated IPs, we can use sqlite3 to store the CNI information in a
database, that allows use to create a well defined data structure that
can be shared between the daemon and the cni plugin with all the "nice"
things that databases  provide.

Performance of adding and deleting 300 interfaces is under 2 seconds:

 cni_test.go:244: 300 success out of 300 in 1.763787501s
--- PASS: TestAddDel (1.76s)
PASS
ok      cni-kindnet     1.787s